Adds amount limit validation

+++ www/components/bet/bet.controller.js
@@ -67,7 +67,10 @@ angular.module('bet')
 								template: data.message
 							});
 
-							BetService.removeInvalidTickets(data.tickets);
+							if(data.tickets) {
+								BetService.removeInvalidTickets(data.tickets);
+							}
+
 							_setBet();
 						}
 					},
